* commit 'bfa03b0dce8038d17660f50e1b9364913390c772':
  Merge remote-tracking branch 'origin/master' into demccccccccj
  样式调整
  Merge remote-tracking branch 'origin/master' into demccccccccj (reverted from commit a9ce2d230d1fd695c8c0b6f313cd9d9171957028)
  merge
  样式调整
  参数对齐

+++ dmestore-service/src/main/java/com/dmeplugin/dmestore/model/VmfsDataInfo.java
@@ -82,7 +82,7 @@ public class VmfsDataInfo {
     /**
      * dme中时延 单位ms
      **/
-    Float latency;
+    Integer latency;
     /**
      * dme中卷ID
      **/
@@ -240,11 +240,11 @@ public class VmfsDataInfo {
         this.writeResponseTime = writeResponseTime;
     }
 
-    public Float getLatency() {
+    public Integer getLatency() {
         return latency;
     }
 
-    public void setLatency(Float latency) {
+    public void setLatency(Integer latency) {
         this.latency = latency;
     }
